db transactions should work (see desciption)
Right now the logic for the db transaction (django transactions are very
nice may I add) is both far too simple (not catching exceptions) as well
as in the completely wrong place. I'm not sure how django structures
usually work but putting logic like this in the controller doesn't sit
right with me.